Update assignments.md
This PR is just a suggestion to what could be improved. I think we need to add an example of creating assignments through the portal. Also, I think we need to make a how to use our product from end-to-end. Starting from creating the custom policy to creating custom assignments and how they intergrate together.